requirements-minimal.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-dotenv==1.2.2
sqlalchemy==2.0.48
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
pydantic[email]==2.12.5
